Release DiFX-2.9.0: Major feature updates and performance improvements

+++ mpifxcorr/src/vdiffile.h
@@ -14,16 +14,6 @@
  *   You should have received a copy of the GNU General Public License     *
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>. *
  ***************************************************************************/
-//===========================================================================
-// SVN properties (DO NOT CHANGE)
-//
-// $Id$
-// $HeadURL: https://svn.atnf.csiro.au/difx/mpifxcorr/trunk/src/mk5.h $
-// $LastChangedRevision$
-// $Author$
-// $LastChangedDate$
-//
-//============================================================================
 #ifndef __VDIFFILE_H__
 #define __VDIFFILE_H__
 
@@ -101,7 +91,7 @@ protected:
   char formatname[64];
 
   unsigned char *readbuffer;
-  int readbuffersize;
+  long long readbuffersize;
   int readbufferslots;
   unsigned int readbufferslotsize;
   int readbufferleftover;
